Update README.md
The scripts-search.py program is, so far, the only code module necessary to complete the job. The problem with the program is it's relatively many command-line switches. 

This program is slated to have a gtk3 gui interface, so that the parameters can be specified on a form, and the output can be either saved or printed. From my perspective, this is a number of months away.